code/requirements.txt
flask==3.0.2
google-generativeai==0.3.2
python-dotenv==1.0.0
setuptools>=41.0.0
llama-index==0.9.27
sentence-transformers==2.5.1
torch>=2.0.0
transformers>=4.36.0
sqlalchemy
psycopg2-binary